CLI mostly working

+++ lib/theme.sh
@@ -1,8 +1,13 @@
 import lib.palette
 
 blck.theme.set() {
-    declare -g -A blck_config
     tpath="$BLCK_HOME/themes/$1.zsh-theme"
+
+    if [ ! -f $tpath ]; then
+        return 1
+    fi
+
+    declare -g -A blck_config
     source $tpath
     blck.theme.copy-theme-var
     blck.theme.unset-theme-var
@@ -18,6 +23,10 @@ blck.theme.list() {
 }
 
 blck.theme.install() {
+    if [ ! -f $1 ]; then
+        return 1
+    fi
+
     cp $1 "$BLCK_HOME/themes/$(basename $1 '.zsh-theme').zsh-theme"
 }
 
@@ -41,3 +50,62 @@ blck.theme.copy-theme-var() {
 blck.theme.unset-theme-var() {
     unset hook_on_load palettes left_prompt right_prompt bottom_left_prompt other_prompts blck_config hooks_after_resize hooks_before_prompt hooks_before_exec hooks_before_accept
 }
+
+blck.theme.print_cmd_list () {
+    echo "Usage: theme set <theme>
+             theme list
+             theme install <path>"
+}
+
+blck.theme.dispatcher() {
+    if [ $# -lt 1 ]; then
+        echo "Expecting one action"
+        blck.theme.print_cmd_list
+        return
+    fi;
+
+    case $1 in
+        'set')  
+                if [ $# -ne 2 ]; then
+                    echo "theme set: Expecting three arguments"
+                    blck.theme.print_cmd_list
+                    return
+                fi
+                
+                blck.theme.set "$2"
+                if [ $? -eq 0 ]; then
+                    echo "Changed theme to $2"
+                else
+                    echo "Theme not found, check list of available themes with \`theme list\`"
+                fi
+                ;;
+        'list')
+                blck.theme.list
+                ;;
+        'install') 
+                if [ $# -ne 2 ]; then
+                    echo "theme install: Expecting three arguments"
+                    blck.theme.print_cmd_list
+                    return
+                fi
+
+                blck.theme.install "$2";
+
+                if [ $? -eq 0 ]; then
+                    echo "Installed theme $(basename $2 '.zsh-theme')\nTry it with \`theme set $(basename $2 '.zsh-theme')\`!"
+                else
+                    echo "Could not find the zsh-theme file \"$2\""
+                fi
+
+                ;;
+        *)
+                echo "Invalid keyword at $2"
+                blck.theme.print_cmd_list
+                ;;
+    esac
+}
+
+# Create CMD Hooks
+echo "REGISTERING"
+blck.cmd.register theme blck.theme.dispatcher
+echo "REGISTERED"
